pyproject.toml
[project]
name = "renting-in-london"
version = "0.1.0"
description = "London rental agent app"
requires-python = ">=3.11"
dependencies = [
  "strands-agents[openai]",
  "strands-agents-tools",
  "httpx",
  "streamlit",
  "pydantic",
  "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
